post fetch by current location radius logic implemented

+++ src/main/java/com/sju18/petmanagement/domain/account/dao/Account.java
@@ -43,5 +43,5 @@ public class Account {
     @Column(nullable = false)
     private Boolean notification;
     @Column(nullable = false)
-    private Integer mapSearchRadius;
+    private Double mapSearchRadius;
 }
